functions/scraper/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.14.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
urllib3==2.1.0
